PR feedback - rename stage-separator/stage-separator to stage-separator/index
