--- packages/compass-aggregations/src/components/stage-separator/index.ts
@@ -1,3 +1,0 @@
-import StageSeparator from './stage-separator';
-export type { StageSeparatorProps } from './stage-separator';
-export default StageSeparator;
