+++ packages/compass-aggregations/src/components/stage-separator/index.tsx
